Attempt to fix cast errors (pace-neutrons/Herbert#283)
* Fix cast errors

+++ _LowLevelCode/cpp/cpp_communicator/serialiser/cpp_serialise.hpp
@@ -11,7 +11,7 @@ enum ser_types{
 
 const uint32_t DIM_MAX = std::numeric_limits<uint32_t>::max();
 
-const double types_size[] = {
+const size_t types_size[] = {
   1,  //   LOGICAL
   1,  //   CHAR
   2,  //   STRING
@@ -156,9 +156,9 @@ struct tag_type {
   unsigned int dim:3;
 };
 
-const double TAG_SIZE = types_size[UINT8];
-const double NELEMS_SIZE = types_size[UINT32];
-const double DIMS_SIZE = types_size[UINT32];
+const size_t TAG_SIZE = types_size[UINT8];
+const size_t NELEMS_SIZE = types_size[UINT32];
+const size_t DIMS_SIZE = types_size[UINT32];
 
 tag_type tag_data(const mxArray* input) {
   int category = mxGetClassID(input);
